+++ models/performance/radostest.py
@@ -8,6 +8,7 @@ SECOND = 1000000    # times are in micro-seconds
 
 import FileStore
 
+
 def kb(val):
     """ number of kilobytes (1024) in a block """
     return val / 1024
@@ -23,14 +24,16 @@ def bw(bs, us):
     return bs / us
 
 
-def radostest(fs, obj_size=16 * MILLION, depth=1, copies=1):
+def radostest(fs, obj_size=16 * MILLION, depth=1, copies=1, crtdlt=False):
     """ compute & display standard filestore test results """
 
-    tc = fs.create(depth=depth)
-    td = fs.delete(depth=depth)
-    print("\t\t     create\t      delete")
-    print("\t\t%6d IOPS\t %6d IOPS" % (iops(tc), iops(td)))
-    print()
+    if crtdlt:
+        tc = fs.create(depth=depth)
+        td = fs.delete(depth=depth)
+        print("\t\t     create\t      delete")
+        print("\t\t%6d IOPS\t %6d IOPS" % (iops(tc), iops(td)))
+        print("")
+
     print("\t    bs\t    rnd read\t   rnd write")
     print("\t -----\t    --------\t   ---------")
     for bs in (4096, 128 * 1024, 4096 * 1024):
